Converted from std::sinf to std::sin

+++ src/texture.h
@@ -38,7 +38,7 @@ class CheckerTexture: public Texture
 
         virtual Color value(float u, float v, const Vec3& p) const override
         {
-            float sines = std::sinf(10*p.x()) * std::sinf(10*p.y()) * std::sinf(10*p.z());
+            float sines = std::sin(10*p.x()) * std::sin(10*p.y()) * std::sin(10*p.z());
 
             if (sines < 0)
                 return odd->value(u, v, p);
